repo reorganization
First of all, HVF is a lot like distribution of an OS along with additional
utilities.  Similarly, z/VM is a collection of 2 operating systems (CP, and
CMS).

So, this commit (1) renames hvf/sys/cp to hvf/sys/shell to show that that
directory contains the user visible shell, and (2) rename hvf/sys to hvf/cp
to indicate that the whole "system" is really just the CP.  NSS live outside
the CP as they should (hvf/nss), and therefore this change only reinforces
the idea that was already there.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>