Load & use symbol table
The loader needs to parse the nucleus ELF headers anyway so let's summarize
the symbol table (.symtab and .strtab) and give a pointer to the start
symbol in cp.  Specifically, it copies over the .symtab and linked .strtab
sections as well as the ELF header at 8KB (real).

Additionally, when we abend, scan the stack frame for addresses which we can
translate to symbol names.  Then we translate them and output them via SCLP.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>