cp: define a FIXME macro that leaves a sclp message
There are far too many fixmes in the code.  Sadly, the compiler simply
discards them.  This usually isn't an issue until one accidentally hits a
"weird" bug which just turns out to be an unhandled (but documented) case in
another part of the code.  Using a macro instead of a comment will let the
compiler string-ify the text, and then at runtime use SCLP to print it out.
This will immediatelly point at problem areas.  So, keep an eye on SCLP from
now on :)

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>